requirements.txt
dominate
six
scipy
opencv_python
lpips
scikit_image
lmdb
tqdm
skimage
visdom
torchmetrics
torch-fidelity